refactor(rust): split authorize.rs and pty.rs under the file-size gate (WI-DP2.5)
The CommandError migration pushed two files past the 300-line limit, and the
gate was already red at HEAD — `pnpm check:all` is frontend-only, so nothing
in the inner loop had reported it.

- browser/refusals.rs — the authorization gate's refusal vocabulary (the five
  typed helpers added in WI-DP2.3). authorize.rs 337 -> 270.
- pty/reader.rs — `pty_start` and its reader thread, longer on its own than
  the other eight PTY commands together. pty.rs 315 -> 197.

`#[tauri::command]` generates a sibling macro that a function-only re-export
does not carry, so command_registry.rs names `pty::reader::pty_start` rather
than re-exporting through pty.rs.

Also applies the rustfmt drift that landed with the earlier WI-DP2.x commits:
the typed signatures grew past 100 columns and `cargo fmt` rewraps them.
check:all never runs `cargo fmt --check`, which is how it reached a commit.

Verified: cargo clippy --all-targets -D warnings (0), cargo test (2080 passed),
pnpm lint:file-size (green).
